UVC-032: Fixed copy paste for zoom
